pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "schedule"
version = "0.1.0"
description = "A Textual-based TUI for quick rescheduling of TaskWarrior tasks via hotkeys"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Matthias Eck", email = "matthias@example.com"}
]
keywords = ["taskwarrior", "cli", "tui", "textual"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "textual>=0.50.0",
    "pyyaml>=6.0",
]

[project.scripts]
schedule = "schedule.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/schedule"]
